fix: csv export for v2 reports
